+++ model/model.go
@@ -116,7 +116,7 @@ func ReadSubmissionComments(app *cview.Application, main *core.MainView, list *c
 		r.UpdateFavoriteStoryAndWriteToDisk(comments)
 		screenWidth := screen.GetTerminalWidth()
 		commentTree := comment.ToString(*comments, config.IndentSize, config.CommentWidth, screenWidth,
-			config.PreserveRightMargin, config.AltIndentBlock)
+			config.PreserveRightMargin, config.AltIndentBlock, config.CommentHighlighting)
 
 		cli.Less(commentTree)
 	})